+++ GB_Vkontakte_AlexeyZhadaev.xcodeproj/xcuserdata/zhadaev.xcuserdatad/xcschemes/xcschememanagement.plist
@@ -7,7 +7,7 @@
 		<key>GB_Vkontakte_AlexeyZhadaev.xcscheme_^#shared#^_</key>
 		<dict>
 			<key>orderHint</key>
-			<integer>2</integer>
+			<integer>4</integer>
 		</dict>
 		<key>Vtentakle_l1_AlexeyZhadaev.xcscheme_^#shared#^_</key>
 		<dict>
