+++ GB_Vkontakte_AlexeyZhadaev/Services/CoreData/GB_Vkontakte_AlexeyZhadaev.xcdatamodeld/GB_Vkontakte_AlexeyZhadaev.xcdatamodel/contents
