+++ Podfile
@@ -8,5 +8,16 @@ target 'GB_Vkontakte_AlexeyZhadaev' do
   pod 'Alamofire', '~> 5.2'
 
   # Pods for GB_Vkontakte_AlexeyZhadaev
-
+	pod 'RealmSwift'
 end
+
+post_install do |installer|
+     installer.pods_project.targets.each do |target|
+         target.build_configurations.each do |config|
+             config.build_settings['EXCLUDED_ARCHS[sdk=iphonesimulator*]'] = 'arm64'
+             config.build_settings['EXCLUDED_ARCHS[sdk=watchsimulator*]'] = 'arm64'
+             config.build_settings['EXCLUDED_ARCHS[sdk=appletvsimulator*]'] = 'arm64'
+    
+         end
+     end
+ end
